Move rebuild_site! helper into separate module
We want to be able to use the `rebuild_site!` helper outside of RSpec
tasks.

This commit moves the method into file `spec/site_helpers.rb`. It is
added to the module `Helpers` so it can be included into the RSpec
configuration using the pattern in the RSpec documentation [[1]]. We
also make it a module function so it can be called as
`Helpers.rebuild_site!` after loading the file.

[1]: https://relishapp.com/rspec/rspec-core/docs/helper-methods/define-helper-methods-in-a-module
