+++ spec/spec_helper.rb
@@ -1,8 +1,8 @@
 $LOAD_PATH.unshift(File.expand_path("../lib", __dir__))
 
 require "govuk_tech_docs"
-require "tempfile"
-require "yaml"
+
+require_relative "site_helper"
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
@@ -23,27 +23,8 @@ require "yaml"
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  def rebuild_site!(config: "config/tech-docs.yml", overrides: {})
-    config_file = Tempfile.new("tech_docs_config")
-
-    begin
-      Dir.chdir("example") do
-        new_config = YAML.load_file(config).merge(overrides)
-        config_file.write(YAML.dump(new_config))
-        config_file.close
-        command = [
-          "rm -rf build",
-          "bundle install --quiet",
-          "CONFIG_FILE=#{config_file.path} NO_CONTRACTS=true middleman build --bail --show-exceptions",
-        ].join(" && ")
-        unless system(command)
-          raise "`middleman build` failed, see the log for more info"
-        end
-      end
-    ensure
-      config_file.unlink
-    end
-  end
+  # Add other helpers
+  config.include Helpers
 
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
